+++ odbc-api/tests/integration.rs
@@ -864,7 +864,7 @@ fn adaptive_columnar_insert_varchar(profile: &Profile) {
         Some(&b"Hello, World!"[..]),
     ];
 
-    let mut buffer = ColumnarRowSet::new(input.len() as u32, iter::once(desc));
+    let mut buffer = ColumnarRowSet::new(input.len(), iter::once(desc));
 
     buffer.set_num_rows(input.len());
     if let AnyColumnViewMut::Text(mut writer) = buffer.column_mut(0) {
@@ -917,7 +917,7 @@ fn adaptive_columnar_insert_varbin(profile: &Profile) {
         Some(&b"Hello, World!"[..]),
     ];
 
-    let mut buffer = ColumnarRowSet::new(input.len() as u32, iter::once(desc));
+    let mut buffer = ColumnarRowSet::new(input.len(), iter::once(desc));
 
     buffer.set_num_rows(input.len());
     if let AnyColumnViewMut::Binary(mut writer) = buffer.column_mut(0) {
